[Q20-4] deprecated interface and moved it
